For adjusted CI configuration.

+++ wcfsetup/install/files/acp/update_com.woltlab.wcf_5.3_invalidateMailForm.php
@@ -5,16 +5,18 @@ use wcf\system\WCF;
 /**
  * Removes the files belonging to the old MailForm in order to make this completely non-functional.
  * 
- * @author	Florian Gail
+ * @author	Tim Duesterhus, Florian Gail
  * @copyright	2001-2020 WoltLab GmbH
  * @license	GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @package	WoltLabSuite\Core
  */
 
-@unlink(WCF_DIR . 'lib/form/MailForm.class.php');
-@unlink(WCF_DIR . 'templates/mail.tpl');
+if (file_exists(WCF_DIR . 'lib/form/MailForm.class.php')) {
+	unlink(WCF_DIR . 'lib/form/MailForm.class.php');
+}
 
-// delete file log entry
+// At this point the MailForm.class.php must be absent. Either the unlink() succeeded or
+// the file already was gone.
 $sql = "DELETE FROM	wcf".WCF_N."_package_installation_file_log
 	WHERE		packageID = ?
 			AND filename = ?";
@@ -23,6 +25,11 @@ $statement->execute([
 	$this->installation->getPackageID(),
 	'lib/form/MailForm.class.php'
 ]);
+
+if (file_exists(WCF_DIR . 'templates/mail.tpl')) {
+	unlink(WCF_DIR . 'templates/mail.tpl');
+}
+
 $sql = "DELETE FROM	wcf".WCF_N."_template
 	WHERE		packageID = ?
 			AND templateName = ?";
